Move the ConfigurationPolicy into the Domain project, so that it can be used by the other WorkerRole projects. Updated tests accordingly
