+++ src/SFA.DAS.EmployerApprenticeshipsService.Web.UnitTests/Infrastructure/StructureMapRegistrationTests/WhenInitalisingTheContainer.cs
@@ -1,5 +1,6 @@
 ﻿using NUnit.Framework;
 using SFA.DAS.EmployerApprenticeshipsService.Domain.Configuration;
+using SFA.DAS.EmployerApprenticeshipsService.Domain.DepedencyResolution;
 using SFA.DAS.EmployerApprenticeshipsService.Web.DependencyResolution;
 using StructureMap;
