+++ src/SFA.DAS.EmployerApprenticeshipsService.Web/DependencyResolution/IoC.cs
@@ -17,6 +17,7 @@
 
 
 using SFA.DAS.EmployerApprenticeshipsService.Domain.Configuration;
+using SFA.DAS.EmployerApprenticeshipsService.Domain.DepedencyResolution;
 
 namespace SFA.DAS.EmployerApprenticeshipsService.Web.DependencyResolution {
     using StructureMap;
